refactor(statsapi)!: rename module to aggregationapi
ServiceNow's official name for the API behind /api/now/stats is the
Aggregate API, and the docs site already labels the module
"Aggregation" — the package name was the remaining inconsistency.
Symbol names (Stats*) and the Now().Stats() accessor are kept, as they
mirror the wire format and URL segment respectively.

Closes #491

BREAKING CHANGE: the statsapi package is renamed; update imports from
github.com/michaeldcanady/servicenow-sdk-go/statsapi to
github.com/michaeldcanady/servicenow-sdk-go/aggregationapi.
