+++ Readme.md
@@ -19,7 +19,7 @@ A Service-Now API client enabling Go programs to interact with Service-Now in a 
 | [Account](accountapi/)                                                | ✔️     |                                                                                                                                                                                                                                                     |
 | [ActivitySubscriptions](actsubapi/)                                   | ✔️     |                                                                                                                                                                                                                                                     |
 | Agent Client Collector                                                | ✖️     |                                                                                                                                                                                                                                                     |
-| [Aggregate (Stats)](statsapi/)                                        | ✔️     |                                                                                                                                                                                                                                                     |
+| [Aggregate (Stats)](aggregationapi/)                                        | ✔️     |                                                                                                                                                                                                                                                     |
 | AI Search External User Mapping                                       | ✖️     |                                                                                                                                                                                                                                                     |
 | Alarm Management Open                                                 | ✖️     |                                                                                                                                                                                                                                                     |
 | [Application Service](appserviceapi/)                                 | ✔️     |                                                                                                                                                                                                                                                     |
