+++ aggregationapi/stats_request_builder.go
@@ -1,4 +1,4 @@
-package statsapi
+package aggregationapi
 
 import (
 	"context"
