+++ aggregationapi/stats_request_builder_get_query_parameters_test.go
@@ -1,4 +1,4 @@
-package statsapi
+package aggregationapi
 
 import (
 	"testing"
