+++ aggregationapi/stats_request_builder_get_request_configuration.go
@@ -1,4 +1,4 @@
-package statsapi
+package aggregationapi
 
 import abstractions "github.com/microsoft/kiota-abstractions-go"
