+++ aggregationapi/stats_result.go
@@ -1,4 +1,4 @@
-package statsapi
+package aggregationapi
 
 import (
 	"github.com/michaeldcanady/servicenow-sdk-go/core"
