+++ website/docs/user-guide/apis/aggregation/aggregate-records.mdx
@@ -36,5 +36,5 @@ func main() {
 
 ## Next steps
 
-- **[Aggregation API reference](../../../apis/stats/index.mdx)**
+- **[Aggregation API reference](../../../apis/aggregation/index.mdx)**
 - **[Build a query](../../query-builder.mdx)**
