+++ website/docs/user-guide/tables/list-and-query-records.mdx
@@ -12,7 +12,7 @@ Fetch the records matching a filter and read fields from each result.
 - You need server-side filtering so you only transfer the records you care
   about
 - You're reading field values out of the results, not just counting them
-  (for counts and aggregates, use the [Aggregation API](../../apis/stats/index.mdx))
+  (for counts and aggregates, use the [Aggregation API](../../apis/aggregation/index.mdx))
 
 ## Required values
